resource delete working properly

+++ routes/resources.js
@@ -16,6 +16,7 @@ var {
   isAdmin,
   isSuperAdmin,
 } = require("../config/auth");
+const resource = require("../models/resource");
 
 router.use(express.urlencoded({ extended: true }));
 
@@ -39,6 +40,7 @@ router.post("/", isAdmin, upload.single("resourceFile"), async (req, res) => {
   var name = req.body.resourceName;
   var subjectCode = req.body.resourceSubjectCode;
   var fileName = req.file.filename;
+  var user=req.user._id;
   await Subject.find({ subjectCode: subjectCode }, async (err, subjects) => {
     if (err) {
       console.log(err);
@@ -47,6 +49,7 @@ router.post("/", isAdmin, upload.single("resourceFile"), async (req, res) => {
         name: name,
         subjectCode: subjectCode,
         fileName: fileName,
+        user:user,
       });
 
       await resource.save();
@@ -132,26 +135,45 @@ router.post("/:id", isAdmin, async (req, res) => {
 });
 
 router.post("/:id/delete", isAdmin, async (req, res) => {
+ 
+  var resourceUser;
+
   Resource.findById(req.params.id, (err, resource) => {
+   
     if (err) {
       console.log(err);
     } else {
-      fs.unlink(`..\\static\\pdf\\${resource.fileName}`, (err) => {
-        if (err) {
-          console.log(err);
-        } else {
-          console.log("File Deleted");
-        }
-      });
+      resourceUser=resource.user;
+      if(req.user._id==resource.user || req.user.isSuperAdmin){
+
+        fs.unlink(`..\\static\\pdf\\${resource.fileName}`, (err) => {
+          if (err) {
+            console.log(err);
+          } else {
+            console.log("File Deleted");
+          }
+        });
+      }
+      // else{
+      //  req.flash('you are not authorized to delete this resource');
+      //   res.redirect(`/resources/${req.params.id}`);
+      // }
     }
-  });
-  Resource.findByIdAndRemove(req.params.id, (err) => {
-    if (err) {
-      console.log(err);
-    } else {
-      res.redirect("/resource");
-    }
-  });
+  }); 
+  if(req.user.isSuperAdmin ||(resourceUser && resourceUser == req.user._id) ){
+
+    Resource.findByIdAndRemove(req.params.id, (err) => {
+      if (err) {
+        console.log(err);
+      } else {
+        res.redirect("/resource");
+      }
+    });
+  }
+  else{
+    req.flash('you are not authorized to delete this resource');
+    res.redirect(`/resources/${req.params.id}`);
+  }
 });
 
 module.exports = router;
